# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the VyoLayer API documentation.",
    "title": "VyoLayer API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:6999",
  "basePath": "/api/v1",
  "paths": {
...
    "/organizations/{orgId}/projects": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Projects"],
        "summary": "List all projects of an organization",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/dto.ProjectDTO"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Projects"],
        "summary": "Create a new project",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "description": "Project to create",
            "name": "project",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateProjectRequestDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/dto.ProjectDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
    "/organizations/{orgId}/projects/{projectId}": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Projects"],
        "summary": "Get a specific project by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Project ID",
            "name": "projectId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.ProjectDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      },
      "put": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Projects"],
        "summary": "Update a specific project by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Project ID",
            "name": "projectId",
            "in": "path",
            "required": true
          },
          {
            "description": "Project to update",
            "name": "project",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.UpdateProjectRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.ProjectDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Projects"],
        "summary": "Delete a specific project by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Project ID",
            "name": "projectId",
            "in": "path",
            "required": true
          },
          {
            "description": "Project to delete",
            "name": "project",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.DeleteProjectRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/organizations/{orgId}/projects/{projectId}/archive": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Projects"],
        "summary": "Archive a specific project by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Project ID",
            "name": "projectId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/organizations/{orgId}/projects/{projectId}/restore": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Projects"],
        "summary": "Restore a specific project by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Project ID",
            "name": "projectId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "dto.CreateProjectRequestDTO": {
      "type": "object",
      "required": ["name"],
      "properties": {
        "description": {
          "type": "string",
          "maxLength": 500,
          "example": "A cool project"
        },
        "name": {
          "type": "string",
          "maxLength": 100,
          "minLength": 3,
          "example": "My Project"
        }
      }
    },
...
    "dto.DeleteProjectRequestDTO": {
      "type": "object",
      "required": ["confirmName"],
      "properties": {
        "confirmName": {
          "type": "string",
          "example": "My Project"
        }
      }
    },
...
    "dto.ProjectDTO": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "createdBy": {
          "type": "string",
          "example": "user_550e8400-e29b-41d4-a716-446655440000"
        },
        "description": {
          "type": "string",
          "example": "A cool project"
        },
        "id": {
          "type": "string",
          "example": "project_550e8400-e29b-41d4-a716-446655440000"
        },
        "isActive": {
          "type": "boolean",
          "example": true
        },
        "maxApiKeys": {
          "type": "integer",
          "example": 5
        },
        "maxMembers": {
          "type": "integer",
          "example": 10
        },
        "memberCount": {
          "type": "integer",
          "example": 3
        },
        "name": {
          "type": "string",
          "example": "My Project"
        },
        "organizationId": {
          "type": "string",
          "example": "org_550e8400-e29b-41d4-a716-446655440000"
        },
        "slug": {
          "type": "string",
          "example": "my-project"
        }
      }
    },
...
    "dto.UpdateProjectRequestDTO": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "maxLength": 500,
          "example": "Updated description"
        },
        "name": {
          "type": "string",
          "maxLength": 100,
          "minLength": 3,
          "example": "My Project"
        }
      }
    },
...
    "response.ErrorDetail": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "example": "ERR_VALIDATION_FAILED"
        },
        "details": {},
        "message": {
          "type": "string",
          "example": "Validation failed"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": true
        },
        "requestId": {
          "type": "string",
          "example": "req-123456"
        },
        "timestamp": {
          "type": "string",
          "example": "2026-02-08T18:30:00Z"
        }
      }
    },
    "response.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/response.ErrorDetail"
        },
        "statusCode": {
          "type": "integer",
          "example": 400
        },
        "success": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "response.Meta": {
      "type": "object",
      "properties": {
        "limit": {
          "type": "integer",
          "example": 10
        },
        "page": {
          "type": "integer",
          "example": 1
        },
        "requestId": {
          "type": "string",
          "example": "req-123456"
        },
        "timestamp": {
          "type": "string",
          "example": "2026-02-08T18:30:00Z"
        },
        "total": {
          "type": "integer",
          "example": 100
        },
        "totalPages": {
          "type": "integer",
          "example": 10
        }
      }
    },
    "response.SuccessResponse": {
      "type": "object",
      "properties": {
        "data": {},
        "message": {
          "type": "string",
          "example": "Operation successful"
        },
        "meta": {
          "$ref": "#/definitions/response.Meta"
        },
        "statusCode": {
          "type": "integer",
          "example": 200
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    }
  }
}